live-transcription-server/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers==3.0.1
accelerate==0.26.0